+++ test/usertype.cpp
@@ -92,6 +92,8 @@ struct non_default_construct {
 };
 
 namespace scn {
+    SCN_CLANG_PUSH
+    SCN_CLANG_IGNORE("-Wpadded")
     template <typename CharT>
     struct scanner<CharT, wrap_default<non_default_construct>>
         : public scanner<CharT, int> {
@@ -107,6 +109,7 @@ namespace scn {
             return {};
         }
     };
+    SCN_CLANG_POP
 }  // namespace scn
 
 TEST_CASE("non_default_construct")